backend/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [ "Marcus Schiesser <mail@marcusschiesser.de>" ]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.12"
fastapi = "^0.109.1"
python-dotenv = "^1.0.0"
llama-index = "0.10.15"
llama-index-core = "0.10.15"
llama-index-agent-openai = "0.1.5"
pygithub = "^2.2.0"
llama-index-readers-github = "^0.1.7"
llama-index-embeddings-huggingface = "^0.1.4"
transformers = "^4.39.1"
torch = "^2.2.1"
sentence-transformers = "^2.6.0"
ipython = "^8.22.2"

[tool.poetry.dependencies.uvicorn]
extras = [ "standard" ]
version = "^0.23.2"

[tool.poetry.dependencies.llama-index-readers-web]
version = "^0.1.6"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"